Add simple panel that displays current date and time.
Preparation steps:

1. Install lemonboys bar
	1.1 git clone https://github.com/LemonBoy/bar.git
	1.2 cd bar
	1.3 make && sudo make install
2. Add the folder where the panel script is to the path
	2.1 So that you can call it just by panel and not the full path
	2.2 See .bashrc changes
	2.3 You may need to reboot for this change to take place
3. Start the panel by calling the script in bspwmrc

Script explanation:

The script sets up a FIFO (first in first out) file for the panel. It
then starts multiple functions that take some information, format it,
and then prints it to stdout on repeat.

These functions are then started and their output is piped into the FIFO
file. Then we set up the panel_bar function to read the fifo file and
output the bar contents based on that. It keeps an internal state that
it updates based on the line reads in the case clauses. Then at the
bottom it takes the gathered infromation, formats it with some lemonbar
formatters, and then prints it to stdout again.

Then at the end we start the panel_bar function and pipe its output in
the lemonbar that then displays it. All of this works because lemonbar
listens to input and redraws the whole panel everytime it gets
something new.

Now that you understand the basics behind the panel script go steal
parts of other peoples configuration!